feat: Implement input handling and settings
This commit introduces the `InputManager` to handle user input during gameplay. It also adds an `InputSettingsView` to configure input mappings and revamps the `ProfileView` to navigate to it.
