+++ Virgo/views/helpers/GameplayPlaybackHelper.swift
@@ -1,0 +1,264 @@
+//
+//  GameplayPlaybackHelper.swift
+//  Virgo
+//
+//  Created by Claude Code on 9/8/2025.
+//
+
+import Foundation
+
+extension GameplayView {
+    // MARK: - Actions
+    func togglePlayback() {
+        isPlaying.toggle()
+        if isPlaying {
+            startPlayback()
+        } else {
+            pausePlayback()
+        }
+    }
+
+    func pausePlayback() {
+        playbackTimer?.invalidate()
+        playbackTimer = nil
+
+        // Stop metronome
+        metronome.stop()
+        
+        // Stop InputManager listening
+        inputManager.stopListening()
+
+        // Save elapsed time when pausing
+        if let startTime = playbackStartTime {
+            pausedElapsedTime += Date().timeIntervalSince(startTime)
+        }
+        playbackStartTime = nil
+
+        bgmPlayer?.pause()
+        Logger.audioPlayback("Paused playback for track: \(track?.title ?? "Unknown")")
+    }
+
+    func startPlayback() {
+        isPlaying = true
+
+        // Start metronome first for timing synchronization
+        guard let track = track else { return }
+        metronome.start(bpm: track.bpm, timeSignature: track.timeSignature)
+
+        // Set start time after metronome starts to ensure sync
+        playbackStartTime = Date()
+        playbackTimer?.invalidate()
+        
+        // Start InputManager listening with song start time
+        if let startTime = playbackStartTime {
+            inputManager.startListening(songStartTime: startTime)
+        }
+
+        // Start BGM playback if available
+        if let bgmPlayer = bgmPlayer {
+            // Check if BGM was previously paused and resume from current position
+            if bgmPlayer.currentTime > 0 && !bgmPlayer.isPlaying {
+                bgmPlayer.play()
+                Logger.audioPlayback("Resumed BGM playback for track: \(track.title)")
+            } else {
+                // Start from beginning
+                bgmPlayer.currentTime = 0
+                bgmPlayer.play()
+                Logger.audioPlayback("Started BGM playbook for track: \(track.title)")
+            }
+        }
+
+        Logger.audioPlayback("Started playback for track: \(track.title)")
+
+        // Initialize playback position
+        currentBeat = 0
+        currentQuarterNotePosition = 0.0
+        totalBeatsElapsed = 0
+        lastBeatUpdate = -1
+        currentBeatPosition = 0.0
+        currentMeasureIndex = 0
+
+        // CRITICAL: Force immediate UI update to show purple bar at position 0 from the start
+        // This ensures the purple bar appears at beat 0 before the first timer update
+        playbackProgress = 0.0
+
+        // Re-enable timer now that MetronomeEngine threading is fixed
+        // Timer must be scheduled on main thread, but we can do the heavy work in background
+        playbackTimer = Timer.scheduledTimer(withTimeInterval: 0.1, repeats: true) { timer in
+            self.updatePlaybackPosition(timer: timer)
+        }
+    }
+
+    func updatePlaybackPosition(timer: Timer) {
+        guard isPlaying else {
+            timer.invalidate()
+            return
+        }
+
+        guard let elapsedTime = calculateElapsedTime(),
+              let track = track else { return }
+
+        let playbackData = calculatePlaybackPosition(elapsedTime: elapsedTime, track: track)
+        updateUIWithPlaybackData(playbackData, timer: timer, track: track)
+    }
+
+    // MARK: - Playback Position Helper Methods
+
+    func calculateElapsedTime() -> Double? {
+        guard let startTime = playbackStartTime else { return nil }
+        let currentSessionTime = Date().timeIntervalSince(startTime)
+        return pausedElapsedTime + currentSessionTime
+    }
+
+    struct PlaybackPositionData {
+        let newBeatIndex: Int
+        let newMeasureIndex: Int
+        let newBeatPosition: Double
+        let newTotalBeats: Int
+        let newProgress: Double
+    }
+
+    func calculatePlaybackPosition(elapsedTime: Double, track: DrumTrack) -> PlaybackPositionData {
+        let secondsPerBeat = 60.0 / Double(track.bpm)
+        let beatsElapsed = elapsedTime / secondsPerBeat
+        let beatsPerMeasure = Double(track.timeSignature.beatsPerMeasure)
+        
+        // Preserve fractional beat values for smooth animation
+        let newMeasureIndex = Int(beatsElapsed / beatsPerMeasure)
+        let beatWithinMeasure = beatsElapsed.truncatingRemainder(dividingBy: beatsPerMeasure)
+        let newBeatPosition = beatWithinMeasure / beatsPerMeasure
+        
+        // Debug logging to trace the calculation
+        if Int(elapsedTime * 10) % 5 == 0 { // Log every 0.5 seconds to avoid spam
+            let logMessage = "Calculation: elapsed=\(elapsedTime), beatsElapsed=\(beatsElapsed), " +
+                           "beatsPerMeasure=\(beatsPerMeasure), beatWithinMeasure=\(beatWithinMeasure), "+
+                           "newBeatPosition=\(newBeatPosition)"
+            Logger.debug(logMessage)
+        }
+        
+        let totalBeats = Int(beatsElapsed)  // Keep for compatibility
+        let newBeatIndex = findClosestBeatIndex(measureIndex: newMeasureIndex, beatPosition: newBeatPosition)
+        let newProgress = min(elapsedTime / cachedTrackDuration, 1.0)
+
+        return PlaybackPositionData(
+            newBeatIndex: newBeatIndex,
+            newMeasureIndex: newMeasureIndex,
+            newBeatPosition: newBeatPosition,
+            newTotalBeats: totalBeats,
+            newProgress: newProgress
+        )
+    }
+
+    func findClosestBeatIndex(measureIndex: Int, beatPosition: Double) -> Int {
+        guard !cachedDrumBeats.isEmpty else { return 0 }
+
+        var left = 0
+        var right = cachedDrumBeats.count - 1
+        var result = 0
+
+        while left <= right {
+            let mid = (left + right) / 2
+            let currentTimePosition = Double(measureIndex) + beatPosition
+            if cachedDrumBeats[mid].timePosition <= currentTimePosition {
+                result = mid
+                left = mid + 1
+            } else {
+                right = mid - 1
+            }
+        }
+
+        return result
+    }
+
+    func updateUIWithPlaybackData(_ data: PlaybackPositionData, timer: Timer, track: DrumTrack) {
+        let shouldUpdate = shouldUpdateUI(with: data)
+        guard shouldUpdate else { return }
+
+        applyUIUpdates(with: data)
+
+        if data.newProgress >= 1.0 {
+            handlePlaybackCompletion(timer: timer, track: track)
+        }
+    }
+
+    func shouldUpdateUI(with data: PlaybackPositionData) -> Bool {
+        return data.newBeatIndex != currentBeat ||
+               data.newBeatPosition != currentBeatPosition ||
+               data.newMeasureIndex != currentMeasureIndex ||
+               data.newTotalBeats != totalBeatsElapsed ||
+               abs(playbackProgress - data.newProgress) > 0.02
+    }
+
+    func applyUIUpdates(with data: PlaybackPositionData) {
+        if data.newBeatIndex != currentBeat {
+            currentBeat = data.newBeatIndex
+        }
+
+        if data.newBeatPosition != currentBeatPosition || data.newMeasureIndex != currentMeasureIndex {
+            currentBeatPosition = data.newBeatPosition
+            currentMeasureIndex = data.newMeasureIndex
+            let beatsPerMeasure = track?.timeSignature.beatsPerMeasure ?? 4
+            let logMessage = "Purple bar update: measure=\(data.newMeasureIndex), " +
+                           "beatPos=\(data.newBeatPosition), " +
+                           "timeSignature=\(track?.timeSignature.displayName ?? "unknown"), " +
+                           "beatsPerMeasure=\(beatsPerMeasure)"
+            Logger.debug(logMessage)
+        }
+
+        if data.newTotalBeats != totalBeatsElapsed {
+            totalBeatsElapsed = data.newTotalBeats
+            currentQuarterNotePosition = Double(data.newMeasureIndex) + data.newBeatPosition
+        }
+
+        if abs(playbackProgress - data.newProgress) > 0.02 {
+            playbackProgress = data.newProgress
+        }
+    }
+
+    func handlePlaybackCompletion(timer: Timer, track: DrumTrack) {
+        timer.invalidate()
+        isPlaying = false
+        metronome.stop()
+        playbackProgress = 0.0
+        currentBeat = 0
+        currentQuarterNotePosition = 0.0
+        totalBeatsElapsed = 0
+        lastBeatUpdate = -1
+        currentBeatPosition = 0.0
+        currentMeasureIndex = 0
+        playbackStartTime = nil
+        pausedElapsedTime = 0.0
+        bgmPlayer?.stop()
+        Logger.audioPlayback("Playback finished for track: \(track.title)")
+    }
+
+    func restartPlayback() {
+        playbackProgress = 0.0
+        currentBeat = 0
+        currentQuarterNotePosition = 0.0
+        totalBeatsElapsed = 0
+        lastBeatUpdate = -1
+        currentBeatPosition = 0.0
+        currentMeasureIndex = 0
+        playbackStartTime = nil
+        pausedElapsedTime = 0.0  // Reset paused time on restart
+        metronome.stop()
+        bgmPlayer?.stop()
+        bgmPlayer?.currentTime = 0
+        Logger.audioPlayback("Restarted playback for track: \(track?.title ?? "Unknown")")
+        if isPlaying {
+            startPlayback()
+        }
+    }
+
+    func skipToEnd() {
+        playbackProgress = 1.0
+        isPlaying = false
+        playbackTimer?.invalidate()
+        playbackTimer = nil
+        playbackStartTime = nil
+        pausedElapsedTime = 0.0  // Reset paused time on skip to end
+        bgmPlayer?.stop()
+        Logger.audioPlayback("Skipped to end for track: \(track?.title ?? "Unknown")")
+    }
+}
