+++ Virgo/views/subviews/KeyCapturingOverlay.swift
@@ -1,4 +1,3 @@
-
 //
 //  KeyCapturingOverlay.swift
 //  Virgo
